Refactor to move HeaderLayout and HamburgerMenu inside header as fragement
