--- src/components/hamburger-menu/index.ts
@@ -1,1 +1,0 @@
-export { HamburgerMenu } from "./HamburgerMenu";
