+++ src/components/header/Header.tsx
@@ -1,35 +1,31 @@
-import React, { FC } from "react";
-import { HamburgerSVG } from "@svgs";
-import { Button, ThemeToggle } from "@components";
-import { NavButtons } from "./components";
-import { scrollToTop } from "@helpers";
-import "./Header.css";
 import { Section } from "@types";
+import React, { useEffect, useState } from "react";
+import { HamburgerMenu, HeaderLayout } from "./components";
+import "./Header.css";
+import { generateNavSections } from "./helpers";
 
-interface HeaderProps {
-  sections: Section[];
-  toggleHamburgerMenu: () => void;
-}
+export const Header = () => {
+  const [sections, setSections] = useState<Section[]>([]);
+  const [isHamburgerOpen, setIsHamburgerOpen] = useState(false);
 
-export const Header: FC<HeaderProps> = ({ toggleHamburgerMenu, sections }) => {
+  const toggleHamburgerMenu = () => {
+    setIsHamburgerOpen(!isHamburgerOpen);
+  };
+
+  useEffect(() => {
+    setSections(generateNavSections());
+  }, []);
+
+  const props = {
+    isHamburgerOpen,
+    sections,
+    onClick: toggleHamburgerMenu,
+  };
+
   return (
-    <header>
-      <div className="header-content">
-        <Button
-          svg={<HamburgerSVG />}
-          ariaLabel="Open Menu"
-          className="hamburger"
-          onClick={toggleHamburgerMenu}
-        />
-        <Button
-          text="Sam McCathie"
-          ariaLabel="Navigate to top of page"
-          className="name-nav"
-          onClick={scrollToTop}
-        />
-        <NavButtons sections={sections} />
-        <ThemeToggle />
-      </div>
-    </header>
+    <>
+      <HeaderLayout {...props} />
+      <HamburgerMenu {...props} />
+    </>
   );
 };
