+++ src/components/header/components/header-child-props.ts
@@ -1,0 +1,7 @@
+import { Section } from "@types";
+
+export interface HeaderChildProps {
+  sections: Section[];
+  onClick: () => void;
+  isHamburgerOpen: boolean;
+}
