+++ src/components/header/components/index.ts
@@ -1,1 +1,4 @@
 export { NavButtons } from "./NavButtons";
+export { HamburgerMenu } from "./HamburgerMenu";
+export { ThemeToggle } from "./ThemeToggle";
+export { HeaderLayout } from "./HeaderLayout";
