+++ src/components/header/helpers/generateNavSections.ts
