+++ src/components/header/helpers/index.ts
@@ -1,1 +1,2 @@
 export { handleScroll } from "./handleScroll";
+export { generateNavSections } from "./generateNavSections";
