--- src/components/theme-toggle/index.ts
@@ -1,1 +1,0 @@
-export { ThemeToggle } from "./ThemeToggle";
