+++ src/helpers/index.ts
@@ -1,2 +1,1 @@
 export { scrollToTop } from "./scrollToTop";
-export { generateNavSections } from "./generateNavSections";
